// src/app/enum/step.enum.ts
export enum StepEnum {
    INIT = "INIT",
    CHOOSE_NAME = "CHOOSE NAME",
    CAPITAINE = "CAPITAINE",
    CLOSE_EYES = "CLOSE EYES",
    VOTE_CAPITAINE = "VOTE CAPITAINE",
    EQUIPAGE = "EQUIPAGE",
    VOTE_EQUIPAGE = "VOTE EQUIPAGE",
    VOTE_VOYAGE = "VOTE VOYAGE",
    PARTY_REVEAL = "PARTY REVEAL",
    FINAL_REVEAL = "FINAL REVEAL"


}